feat: Modification of the CreateProfileModal and ProfileForm components for the creation and editing of profiles.
